Fix typo (Conditionnable -> Conditionable)
