+++ src/Glpi/Form/Comment.php
@@ -37,8 +37,8 @@ namespace Glpi\Form;
 
 use CommonDBChild;
 use Glpi\Application\View\TemplateRenderer;
-use Glpi\Form\Condition\ConditionnableVisibilityInterface;
-use Glpi\Form\Condition\ConditionnableVisibilityTrait;
+use Glpi\Form\Condition\ConditionableVisibilityInterface;
+use Glpi\Form\Condition\ConditionableVisibilityTrait;
 use Log;
 use Override;
 use Ramsey\Uuid\Uuid;
@@ -47,9 +47,9 @@ use RuntimeException;
 /**
  * Comment of a given helpdesk form's section
  */
-final class Comment extends CommonDBChild implements BlockInterface, ConditionnableVisibilityInterface
+final class Comment extends CommonDBChild implements BlockInterface, ConditionableVisibilityInterface
 {
-    use ConditionnableVisibilityTrait;
+    use ConditionableVisibilityTrait;
 
     public static $itemtype = Section::class;
     public static $items_id = 'forms_sections_id';
